requirements.txt
pymongo~=4.10.1
boto3~=1.35.59
requests~=2.32.3
pandas~=2.2.3
setuptools~=75.6.0
matplotlib~=3.10.0
pyyaml